Use version manager
